Update tests to use up-to-date neutron
The tests were referencing some things in neutron that have been moved or
removed.  The policy engine used to be in neutron.openstack.common.policy but
has since been moved to neutron.policy.

The neutron policy also used to have a method get_admin_roles, but it has since
been removed.  Also, the base upstream neutron context would use that
policy.get_admin_roles() method to inject the admin roles into the
context.roles attribute.  This is no longer the case and a test had to be
removed that was testing for this specific behavior and another test had
to be modified to not expect it.  [1] is the commit upstream where these
changes were made.

[1] https://github.com/openstack/neutron/commit/734e77365b0f241a3cea0f3c9dfb1d5fcf6eac8c