Set advsvc for the neutron context based on policy
JIRA:NCP-1819

Neutron's context has an attribute advsvc that is meant to evaluate to true
if an advanced service (lbaas, fwaas, vpnaas) is using the context.  This will
be used for lbaas at first so that quark can check if requestor is an
advanced service and allow it to do certain things.  This will be done in a
patch to quark though.

This also adds a mock.patch.stopall() cleanup method to the base tests's
setUp method so that every patcher that is created does not have to be stopped.