Add policy check for is_admin post context creation
JIRA:NCP-2099

If a user is not an admin, because tenant_id and user_id exist, is_admin
will be False on context creation. Check policy to see if the current
user has the proper role to have privileges elevated to admin. Some
cleanup of tests_neutron_context was made for unused values.